+++ src/main/resources/static/assets/admin/js/product.js
@@ -156,16 +156,21 @@ app.controller("ctrl-product", function($scope, $http) {
 
 	$scope.delete = function(id) {
 		var url = `${pathProduct}/product/${id}`;
-		$http.delete(url).then(resp => {
-			// tìm ra phần tử tại vị trí sẽ xóa.
-			var index = $scope.items.findIndex(item => item.id == id);
-			$scope.items.splice(index, 1); // tại vị trí đó và xóa 1 phần tử
-			$scope.reset();
-			console.log("Success", resp);
-			alert("Xóa sản phẩm thành công");
-		}).catch(error => {
-			console.log("lỗi xoá", error);
-		});
+		var xacnhan = confirm('Bạn có chắc muốn xóa!');
+		if(xacnhan) {
+			$http.delete(url).then(resp => {
+				// tìm ra phần tử tại vị trí sẽ xóa.
+				var index = $scope.items.findIndex(item => item.id == id);
+				$scope.items.splice(index, 1); // tại vị trí đó và xóa 1 phần tử
+				$scope.reset();
+				console.log("Success", resp);
+				alert("Xóa sản phẩm thành công");
+			}).catch(error => {
+				console.log("lỗi xoá", error);
+			});	
+		}else{
+			console.log("Cancel");
+		}
 	}
 
 	// Upload img
